requirements.txt
Flask==2.0.1
requests==2.22.0
paho-mqtt==1.5.1
fritzconnection==1.5.0
